rename of lockboxclient to contractclient
