+++ cmd/neard/contractclient/broker.go
@@ -1,4 +1,4 @@
-package lockboxclient
+package contractclient
 
 import (
 	"context"
